changed some naming and changed routing to not use queries
